Reworked File package to Output. Renamed those files to be prefixed with output. Modulized to their job. PictureOutput is entry point to system of moving or copying files into the root which is given.
Created a PictureFile that acts as a wrapper over the path and file and contains the imageRatios for now.

Reworked it so that for now we can just grab all the wallpapers in a given directory and dump them into a root folder.

changed packages to be DarkMage530.BulkImager... instead of just Main

Next:
 should be getting ratings for these wallpapers
 organize directory as rating / resolution