Finished up the rough run for iqd parsing, getting an iqdbImage and setting it to PictureFile. Can also ask IqdbImage for things like AverageRating, HarshestRating, BestRating etc... Added a unit test for the IqdbImage ratings.
We could now do a run through P to get wallpapers although it might be really slow. Would like to make this multithreaded to an extent. So anything thats not a Wallpaper keeps going and we just keep queueing up the wallpapers to be processed concurrently.

We can do a work settings one that uses HarshestRating instead of the Average.

Modify the driver so that we can give it certain "commands" on start up to basically configure what happens.

For instance when building Work wallpapers we'd use the HarshestRating. But for Home Best or Average is probably satisfactory.